[CommunicationTokenCredential] Rename all the references of CommunicationUserCredential to CommunicationTokenCredential
