+++ sdk/communication/AzureCommunication/Source/Authentication/CommunicationTokenCredentialProtocol.swift
@@ -29,32 +29,13 @@ import AzureCore
 #endif
 import Foundation
 /**
- The Azure Communication Services User token credential. 
+ Protocol defining the shape of credentials used with Azure Communication Services.
  */
-internal class StaticUserCredential: CommunicationTokenCredentialProtocol {
-    private let accessToken: CommunicationAccessToken
-
-    /**
-     Creates a static `CommunicationTokenCredential` object from the provided token.
-        
-     - Parameter token: The static token to use for authenticating all requests.
-     
-     - Throws: `AzureError` if the provided token is not a valid token.
-     
-     - SeeAlso: ` CommunicationTokenCredential.init(...)`
-     */
-    public init(token: String) throws {
-        self.accessToken = try JwtTokenParser.createAccessToken(token)
-    }
-
+public protocol CommunicationTokenCredentialProtocol {
     /**
      Retrieve an access token from the credential.
-     
      - Parameter completionHandler: Closure that accepts an optional `AccessToken` or optional `Error` as parameters.
      `AccessToken` returns a token and an expiry date if applicable. `Error` returns `nil` if the current token can be returned.
-
      */
-    public func token(completionHandler: AccessTokenRefreshOnCompletion) {
-        completionHandler(accessToken, nil)
-    }
+    func token(completionHandler: @escaping AccessTokenRefreshOnCompletion)
 }
