+++ sdk/communication/AzureCommunication/Tests/ObjCTokenParserTests.m
@@ -36,7 +36,7 @@
 - (void)xtest_ObjCThrowsIfInvalidTokenFoo {
     NSString *invalidFoo = @"foo";
     NSError *error = nil;
-    CommunicationUserCredential *credential = [[CommunicationUserCredential alloc] initWithToken:invalidFoo
+    CommunicationTokenCredential *credential = [[CommunicationTokenCredential alloc] initWithToken:invalidFoo
                                                                                            error:&error];
     XCTAssertNil(credential);
     XCTAssertNotNil(error);
@@ -45,7 +45,7 @@
 - (void)xtest_ObjCThrowsIfInvalidTokenFormat {
     NSString *invalidToken = @"foo.bar.foobar";
     NSError *error = nil;
-    CommunicationUserCredential *credential = [[CommunicationUserCredential alloc] initWithToken:invalidToken
+    CommunicationTokenCredential *credential = [[CommunicationTokenCredential alloc] initWithToken:invalidToken
                                                                                            error:&error];
     XCTAssertNil(credential);
     XCTAssertNotNil(error);
@@ -54,7 +54,7 @@
 - (void)xtest_ObjCThrowsWhenInitWithBlock {
     NSString *invalidToken = @"foo.bar";
     NSError *error = nil;
-    CommunicationUserCredential *credential = [[CommunicationUserCredential alloc] initWithInitialToken:invalidToken
+    CommunicationTokenCredential *credential = [[CommunicationTokenCredential alloc] initWithInitialToken:invalidToken
                                                                                      refreshProactively:YES
                                                                                                   error:&error
                                                                                          tokenRefresher:^(void (^ _Nonnull block)
