+++ sdk/communication/AzureCommunicationChat/README.md
@@ -107,7 +107,7 @@ guard let baseUrl = URL(string: "https://<resource>.communication.azure.com") el
 }
 
 let authPolicy = try CommunicationUserCredentialPolicy(
-    credential: credential ?? CommunicationUserCredential(token: "<user_access_token>")
+    credential: credential ?? CommunicationTokenCredential(token: "<user_access_token>")
 )
 let options = AzureCommunicationChatClientOptions(
     logger: ClientLoggers.default,
